# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: CleanOps Pro - Multi-tenant SaaS API for cleaning businesses
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/{id}/scorecards:
    get:
      operationId: getUserScorecards
      tags: [users]
      summary: Get employee scorecards
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Employee scorecards
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScorecardListResponse"
...
  /timeclock:
    get:
      operationId: listTimeclock
      tags: [timeclock]
      summary: List time clock entries
      security:
        - bearerAuth: []
      parameters:
        - name: user_id
          in: query
          schema:
            type: integer
        - name: job_id
          in: query
          schema:
            type: integer
        - name: flagged
          in: query
          schema:
            type: boolean
        - name: date_from
          in: query
          schema:
            type: string
            format: date
        - name: date_to
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Time clock entries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimeclockListResponse"
  /timeclock/clock-in:
    post:
      operationId: clockIn
      tags: [timeclock]
      summary: Clock in to a job
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClockInRequest"
      responses:
        "201":
          description: Clock in recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimeclockEntry"
  /timeclock/{id}/clock-out:
    post:
      operationId: clockOut
      tags: [timeclock]
      summary: Clock out from a job
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClockOutRequest"
      responses:
        "200":
          description: Clock out recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimeclockEntry"
...
  /scorecards:
    get:
      operationId: listScorecards
      tags: [scorecards]
      summary: List scorecards
      security:
        - bearerAuth: []
      parameters:
        - name: user_id
          in: query
          schema:
            type: integer
        - name: client_id
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of scorecards
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScorecardListResponse"

    post:
      operationId: createScorecard
      tags: [scorecards]
      summary: Create scorecard
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateScorecardRequest"
      responses:
        "201":
          description: Scorecard created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scorecard"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: integer
...
  schemas:
...
    TimeclockEntry:
      type: object
      properties:
        id:
          type: integer
        job_id:
          type: integer
        user_id:
          type: integer
        user_name:
          type: string
        clock_in_at:
          type: string
          format: date-time
        clock_out_at:
          type: string
          format: date-time
          nullable: true
        clock_in_lat:
          type: number
          nullable: true
        clock_in_lng:
          type: number
          nullable: true
        clock_out_lat:
          type: number
          nullable: true
        clock_out_lng:
          type: number
          nullable: true
        distance_from_job_ft:
          type: number
          nullable: true
        flagged:
          type: boolean
        duration_hours:
          type: number
          nullable: true
      required:
        - id
        - job_id
        - user_id
        - user_name
        - clock_in_at
        - flagged
    TimeclockListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/TimeclockEntry"
        total:
          type: integer
      required:
        - data
        - total
    ClockInRequest:
      type: object
      properties:
        job_id:
          type: integer
        lat:
          type: number
          nullable: true
        lng:
          type: number
          nullable: true
      required:
        - job_id
    ClockOutRequest:
      type: object
      properties:
        lat:
          type: number
          nullable: true
        lng:
          type: number
          nullable: true
...
    Scorecard:
      type: object
      properties:
        id:
          type: integer
        job_id:
          type: integer
        user_id:
          type: integer
        user_name:
          type: string
        client_id:
          type: integer
        client_name:
          type: string
        score:
          type: integer
          minimum: 0
          maximum: 4
        comments:
          type: string
          nullable: true
        excluded:
          type: boolean
        created_at:
          type: string
          format: date-time
      required:
        - id
        - job_id
        - user_id
        - user_name
        - client_id
        - client_name
        - score
        - excluded
        - created_at
    ScorecardListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Scorecard"
        total:
          type: integer
        average_score:
          type: number
          nullable: true
      required:
        - data
        - total
    CreateScorecardRequest:
      type: object
      properties:
        job_id:
          type: integer
        user_id:
          type: integer
        client_id:
          type: integer
        score:
          type: integer
          minimum: 0
          maximum: 4
        comments:
          type: string
          nullable: true
        excluded:
          type: boolean
      required:
        - job_id
        - user_id
        - client_id
        - score
...